% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{cars}
\format{
A data frame with 10 rows and 2 columns:
\describe{
  \item{x1}{first var}
  \item{x2}{second var}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
cars
}
\description{
A complete data set of measurements of cars
}
\keyword{datasets}
