% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboxcox.R
\name{myboxcox}
\alias{myboxcox}
\title{myboxcox}
\usage{
myboxcox(x)
}
\arguments{
\item{x}{vector}
}
\value{
a value for the MLE and a plot
}
\description{
Returns the MLE for lambda
}
\examples{
\dontrun{myboxcox(x)}
}
