% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propnorm.R
\name{propnorm}
\alias{propnorm}
\title{propnorm}
\usage{
propnorm(df, varfact, ...)
}
\arguments{
\item{df}{data frame}

\item{varfact}{list of variable names that are factors.}

\item{...}{passes additional arguments to the function}
}
\value{
a plot and a named list
}
\description{
A function which tests normality for each variable and the data as multivariate
}
\examples{
\dontrun{propnorm(t46, facts)}
}
