% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statdist.R
\name{statdist}
\alias{statdist}
\title{statdist}
\usage{
statdist(df, contour = 0.05)
}
\arguments{
\item{df}{a data frame}

\item{contour}{level of contour (default is 0.05)}
}
\value{
a Chi-Sq plot and a named list
}
\description{
A function which tests for multivariate normality
}
\examples{
\dontrun{statdist(cars)}
}
