% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t41.R
\docType{data}
\name{t41}
\alias{t41}
\title{Table 4-1}
\format{
A data frame with 42 rows and 1 column:
\describe{
  \item{Radiation}{Radiation Measurement}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
t41
}
\description{
A complete data set of 42 measurements of Radiation
}
\keyword{datasets}
