% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t46.R
\docType{data}
\name{t46}
\alias{t46}
\title{t46}
\format{
A data frame with 130 rows and 7 columns:
\describe{
  \item{Indep}{Independent}
  \item{Supp}{Supplement}
  \item{Benev}{Benevolence}
  \item{Conform}{Conform}
  \item{Leader}{Leadership}
  \item{Gender}{Gender}
  \item{Socio}{Socio-Econ Status}
  ...
}
}
\source{
<https://www.amazon.com/Applied-Multivariate-Statistical-Analysis-6th/dp/0131877151>
}
\usage{
t46
}
\description{
A complete data set of measurements
}
\keyword{datasets}
